Preparational refactor in orionldMhdConnectionInit+Treat
