+++ test/functionalTest/cases/0000_ngsild/ngsild_error_reject_entity_if_jsonld_link_header_is_provided_with_jsonld_payload.test
@@ -49,7 +49,7 @@ echo
 01. Attempt to create an entity with application/ld+json and with context in HTTP Link Header - see error
 =========================================================================================================
 HTTP/1.1 400 Bad Request
-Content-Length: 233
+Content-Length: 219
 Content-Type: application/json
 Date: REGEX(.*)
